refactor(profile): factorized code from both popup
Reduces massive code duplication between comment and post popup
